Testing!

+++ backend/Backend/Models/Spot.cs
@@ -1,8 +1,5 @@
 ﻿namespace GuelphNavigator.Backend.Models;
 
-public record Spot(string Id, string Name, string PhotosphereUrl, int PositionX, int PositionY, SpotConnection[] connections);
+public record Spot(string Id, string Name, string PhotosphereUrl, double PositionX, double PositionY, SpotConnection[] connections);
 
 public record SpotConnection(Spot EndSpot, int Weight);
-
-//public record SpotPath(double length, );
-
